requirements.txt
discord.py
yt-dlp
Pillow
python-dotenv
aiohttp

